Create safe num_updates access on retrying failed training
Summary:
The num_updates reference is unsafe since it is only defined in the for-loop without any assertion that the loop will execute at least once. We could either try to add this assertion to make it more clear as to what is happening, or we could set num_updates to some default value if the loop is never executed.

I'm open to doing either one.

Reviewed By: chtran

Differential Revision: D22598029

fbshipit-source-id: 4d20b6a415a783fe8c450f3998b7767861da8c06